refactor(polecat): extract swarm/watch/analyze/reset_stalled/summary into modules (Step 1 of Polecat v2 epic-4234682b)
Pure mechanical extraction — no behavior change. The four remaining
sub-commands (analyze, reset-stalled, watch, summary) move out of
polecat/cli.py into dedicated modules; swarm.py was already extracted
in a prior change and is left untouched.

- polecat/diagnostics.py — analyze + reset-stalled (and their imports)
- polecat/watch.py       — watch command + private _send_notification
- polecat/summary.py     — summary command + private parse_duration
- polecat/cli.py         — imports each command and registers via main.add_command

Helpers used only by a single extracted command (parse_duration,
_send_notification) move with their command. Helpers used by multiple
commands stay in cli.py. cli.py shrinks from 5575 to 4784 lines.

Verified:
- polecat <subcmd> --help is byte-identical pre/post for all 5 subcmds
- pytest tests/polecat/ — 348 passed
- pytest tests/        — 1464 passed, 37 skipped
- pre-commit run --all-files — passes
