+++ polecat/cli.py
@@ -30,7 +30,6 @@ if str(REPO_ROOT / "aops-core") not in sys.path:
 import click
 from lib.agent_env import apply_env_mappings, get_container_env_forwards
 from manager import PolecatManager
-from observability import metrics
 from validation import TaskIDValidationError, validate_task_id_or_raise
 
 # Turn budget for headless Claude runs.
@@ -4719,545 +4718,23 @@ def run(
         print(f"   To finish manually: cd {worktree_path} && polecat finish")
 
 
-@main.command()
-@click.argument("task_id")
-@click.option("--transcript-lines", "-n", default=20, help="Number of transcript lines to show")
-@click.pass_context
-def analyze(ctx, task_id, transcript_lines):
-    """Diagnose a stalled or failed task.
-
-    Shows task metadata, worktree status, transcript tail, and suggested
-    remediation actions for tasks that are stuck in_progress.
-
-    Examples:
-        polecat analyze aops-abc12345     # Full diagnostic
-        polecat analyze aops-abc12345 -n 50  # Show more transcript
-    """
-    from datetime import datetime
-
-    # Validate task ID
-    try:
-        validate_task_id_or_raise(task_id)
-    except TaskIDValidationError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    manager = PolecatManager(home_dir=ctx.obj.get("home"))
-
-    # Load task
-    if manager.storage is not None:
-        task = manager.storage.get_task(task_id)
-    else:
-        from polecat.pkb_bridge import get_task as pkb_get_task
-
-        task = pkb_get_task(task_id)
-    if not task:
-        print(f"❌ Task not found: {task_id}", file=sys.stderr)
-        sys.exit(1)
-
-    print(f"🔍 Analyzing task: {task_id}")
-    print("=" * 60)
-
-    # --- Section 1: Task Metadata ---
-    print("\n📋 TASK METADATA")
-    print(f"   Title:    {task.title}")
-    print(f"   Status:   {task.status}")
-    print(f"   Assignee: {task.assignee or '(none)'}")
-    print(f"   Project:  {task.project or 'aops'}")
-    print(f"   Priority: P{task.priority}")
-
-    # Calculate staleness
-    if task.modified:
-        now = datetime.now().astimezone()
-        modified = task.modified
-        if modified.tzinfo is None:
-            modified = modified.replace(tzinfo=UTC)
-        age = now - modified
-        hours = age.total_seconds() / 3600
-        print(f"   Modified: {modified.isoformat()} ({hours:.1f}h ago)")
-
-        # Flag staleness
-        if hours > 4:
-            print(f"   ⚠️  STALE: No activity for {hours:.1f} hours")
-
-    # --- Section 2: Worktree Status ---
-    print("\n📁 WORKTREE STATUS")
-    worktree_path = manager.polecats_dir / task_id
-
-    if not worktree_path.exists():
-        print(f"   ❌ Worktree not found at {worktree_path}")
-        print("   💡 Suggestion: Task may not have been started, or worktree was nuked")
-    else:
-        print(f"   ✓ Worktree exists at {worktree_path}")
-
-        # Check git status
-        import subprocess
-
-        git_status = subprocess.run(
-            ["git", "status", "--porcelain"],
-            cwd=worktree_path,
-            capture_output=True,
-            text=True,
-            check=False,
-        )
-
-        if git_status.returncode == 0:
-            if git_status.stdout.strip():
-                changes = git_status.stdout.strip().split("\n")
-                print(f"   ⚠️  Uncommitted changes ({len(changes)} files):")
-                for line in changes[:5]:
-                    print(f"      {line}")
-                if len(changes) > 5:
-                    print(f"      ... and {len(changes) - 5} more")
-            else:
-                print("   ✓ Working tree clean")
-        else:
-            print(f"   ❌ Git status failed: {git_status.stderr.strip()}")
-
-        # Check branch and commits
-        branch_result = subprocess.run(
-            ["git", "rev-parse", "--abbrev-ref", "HEAD"],
-            cwd=worktree_path,
-            capture_output=True,
-            text=True,
-            check=False,
-        )
-        if branch_result.returncode == 0:
-            branch = branch_result.stdout.strip()
-            print(f"   Branch: {branch}")
-
-            # Check commits ahead of main
-            commits_result = subprocess.run(
-                ["git", "log", "--oneline", "origin/main..HEAD"],
-                cwd=worktree_path,
-                capture_output=True,
-                text=True,
-                check=False,
-            )
-            if commits_result.returncode == 0 and commits_result.stdout.strip():
-                commits = commits_result.stdout.strip().split("\n")
-                print(f"   Commits ahead of main ({len(commits)}):")
-                for commit in commits[:3]:
-                    print(f"      {commit}")
-                if len(commits) > 3:
-                    print(f"      ... and {len(commits) - 3} more")
-            else:
-                print("   No commits ahead of main")
-
-    # --- Section 3: Transcript (if available) ---
-    print("\n📜 TRANSCRIPT")
-    try:
-        from lib.paths import find_polecat_transcript
-
-        transcript_path = find_polecat_transcript(task_id)
-    except ImportError:
-        transcript_path = manager.home_dir / "transcripts" / f"{task_id}.jsonl"
-
-    if not transcript_path.exists():
-        print(f"   (No transcript found at {transcript_path})")
-        print("   💡 Transcript capture may not be enabled yet")
-    else:
-        import json
-
-        try:
-            # Read last N lines
-            with open(transcript_path) as f:
-                lines = f.readlines()
-
-            if not lines:
-                print("   (Transcript file is empty)")
-            else:
-                print(
-                    f"   Showing last {min(transcript_lines, len(lines))} of {len(lines)} entries:"
-                )
-                print()
-                for line in lines[-transcript_lines:]:
-                    try:
-                        entry = json.loads(line)
-                        # Format depends on transcript structure
-                        if "type" in entry:
-                            print(
-                                f"   [{entry.get('type', '?')}] {entry.get('message', entry.get('content', str(entry)[:80]))}"
-                            )
-                        else:
-                            print(f"   {str(entry)[:100]}")
-                    except json.JSONDecodeError:
-                        print(f"   {line.strip()[:100]}")
-        except Exception as e:
-            print(f"   ❌ Failed to read transcript: {e}")
-
-    # --- Section 4: Suggested Remediation ---
-    print("\n💡 SUGGESTED ACTIONS")
-
-    status_str = task.status or ""
-
-    if status_str == "in_progress":
-        if not worktree_path.exists():
-            print("   1. Task claimed but no worktree - may have crashed during setup")
-            print(
-                f"      → Reset: polecat reset-stalled --hours 0 --project {task.project or 'aops'}"
-            )
-            print("      → Or retry: polecat run -t {task_id}")
-        elif hours > 4:
-            print("   1. Task appears stalled (no activity > 4h)")
-            print("      → Check if agent is still running")
-            print("      → Reset if abandoned: polecat reset-stalled")
-            print(f"      → Or manually finish: cd {worktree_path} && polecat finish")
-        else:
-            print("   1. Task is in progress and appears active")
-            print("      → Wait for agent to complete, or check logs")
-    elif status_str == "merge_ready":
-        print("   1. Task ready to merge")
-        print("      → Run: polecat merge")
-    elif status_str == "review":
-        print("   1. Task needs human review before merging")
-        print(f"      → Review changes: cd {worktree_path}")
-        print("      → Then set status to merge_ready or fix issues")
-    elif status_str == "blocked":
-        print("   1. Task is blocked")
-        print("      → Check task body for blocker details")
-        if task.depends_on:
-            print(f"      → Depends on: {', '.join(task.depends_on)}")
-    elif status_str == "done":
-        print("   1. Task is already complete ✓")
-        if worktree_path.exists():
-            print(f"      → Consider cleanup: polecat nuke {task_id}")
-    else:
-        print(f"   Status is '{status_str}' - no specific suggestions")
-
-    print()
-
-
-@main.command("reset-stalled")
-@click.option("--project", "-p", help="Filter by project")
-@click.option("--hours", default=4.0, help="Hours since last modification (default: 4)")
-@click.option("--dry-run", is_flag=True, help="Show what would be reset without changing")
-@click.option(
-    "--force",
-    "-f",
-    is_flag=True,
-    help="Skip confirmation prompt (required in non-interactive mode)",
-)
-@click.pass_context
-def reset_stalled(ctx, project, hours, dry_run, force):
-    """Reset stalled in_progress tasks back to active.
-
-    Finds tasks that have been in_progress for > N hours and resets them.
-    Useful for cleaning up after crashed/abandoned agents.
-    """
-    from datetime import datetime, timedelta
-
-    _TaskIndex = None
-    try:
-        from lib.task_index import TaskIndex as _TaskIndex
-    except ImportError:
-        pass
-
-    manager = PolecatManager(home_dir=ctx.obj.get("home"))
-
-    # Calculate cutoff time
-    cutoff = datetime.now().astimezone() - timedelta(hours=hours)
-
-    print(f"Checking for tasks stalled since {cutoff.isoformat()}...")
-
-    # List tasks
-    if manager.storage is not None:
-        try:
-            from lib.task_model import TaskStatus
-
-            candidates = manager.storage.list_tasks(status=TaskStatus.IN_PROGRESS, project=project)
-        except ImportError:
-            candidates = []
-    else:
-        from polecat.pkb_bridge import list_tasks as pkb_list_tasks
-
-        candidates = pkb_list_tasks(status="in_progress", project=project)
-
-    stalled = []
-    pr_locked = []
-    for task in candidates:
-        # Ensure timezone awareness
-        task_mod = task.modified
-        if task_mod is None:
-            continue
-        if task_mod.tzinfo is None:
-            task_mod = task_mod.replace(tzinfo=UTC)
-
-        if task_mod < cutoff:
-            # Never reset tasks that have an open PR — they are locked by prior work.
-            if task.pr_url or task.pr:
-                pr_locked.append(task)
-            else:
-                stalled.append(task)
-
-    if pr_locked:
-        print(f"Skipping {len(pr_locked)} PR-locked tasks (have open PR — will not reset):")
-        for t in pr_locked:
-            pr_ref = t.pr_url or f"#{t.pr}"
-            print(f"  [{t.id}] {t.title} (PR: {pr_ref})")
-
-    if not stalled:
-        print("No stalled tasks found.")
-        return
-
-    print(f"Found {len(stalled)} stalled tasks (modified > {hours}h ago):")
-    for t in stalled:
-        print(f"  [{t.id}] {t.title} (modified: {t.modified.isoformat()})")
-
-    if dry_run:
-        print("\nDry run: no changes made.")
-        return
-
-    if not force:
-        print(f"\nError: This will reset {len(stalled)} tasks. Use --force to confirm.")
-        sys.exit(1)
-
-    reset_count = 0
-    for task in stalled:
-        try:
-            task.status = "queued"
-            task.assignee = None
-            if manager.storage is not None:
-                try:
-                    from lib.task_model import TaskStatus
-
-                    _status_attr = getattr(
-                        TaskStatus, "QUEUED", getattr(TaskStatus, "ACTIVE", None)
-                    )
-                    if _status_attr is not None:
-                        task.status = _status_attr.value
-                except (ImportError, AttributeError):
-                    pass
-                manager.storage.save_task(task)
-            else:
-                from polecat.pkb_bridge import save_task as pkb_save
-
-                pkb_save(task)
-            reset_count += 1
-        except Exception as e:
-            print(f"Failed to reset {task.id}: {e}", file=sys.stderr)
-
-    # Rebuild index (only relevant for legacy TaskIndex)
-    if reset_count > 0 and _TaskIndex is not None and manager.storage is not None:
-        try:
-            data_root = manager.storage.data_root
-            index = _TaskIndex(data_root)
-            index.rebuild_fast()
-            print("Index rebuilt.")
-        except Exception as e:
-            print(f"Warning: Failed to rebuild index: {e}", file=sys.stderr)
-
-    print(f"\n✅ Reset {reset_count} tasks.")
-
-
-def _send_notification(title: str, message: str, urgency: str = "normal"):
-    """Send a desktop notification via notify-send if available.
-
-    Args:
-        title: Notification title
-        message: Notification body
-        urgency: low, normal, or critical
-    """
-    import shutil
-
-    print(f"[{urgency.upper()}] {title}: {message}")
-
-    if shutil.which("notify-send"):
-        try:
-            import subprocess
-
-            subprocess.run(
-                ["notify-send", "-u", urgency, title, message],
-                check=False,
-                capture_output=True,
-            )
-        except Exception:
-            pass
-
-
-@main.command()
-@click.option(
-    "--interval",
-    "-i",
-    default=300,
-    help="Polling interval in seconds (default: 300 = 5 min)",
-)
-@click.option(
-    "--stall-threshold",
-    "-s",
-    default=30,
-    help="Minutes without progress before stall alert (default: 30)",
-)
-@click.option("--project", "-p", help="Project to monitor (default: all)")
-@click.pass_context
-def watch(ctx, interval, stall_threshold, project):
-    """Monitor swarm activity and send desktop notifications.
-
-    Runs as a background process that:
-    - Polls for new PRs and merge_ready tasks
-    - Sends notification when a new PR is filed
-    - Alerts if swarm stalls (no progress in threshold minutes)
-
-    Examples:
-        polecat watch              # Default: poll every 5min, stall at 30min
-        polecat watch -i 60        # Poll every 60 seconds
-        polecat watch -s 60        # Alert after 60min of no progress
-        polecat watch &            # Run in background
-    """
-    import signal
-    import time
-    from datetime import datetime, timedelta
-
-    _use_legacy = False
-    try:
-        from lib.task_model import TaskStatus
-
-        _use_legacy = True
-    except ImportError:
-        pass
-
-    manager = PolecatManager(home_dir=ctx.obj.get("home"))
-
-    # Track seen PRs and last activity time
-    seen_merge_ready = set()
-    seen_review = set()
-    last_activity = datetime.now().astimezone()
-
-    # Graceful shutdown
-    stop_requested = False
-
-    def handle_signal(signum, frame):
-        nonlocal stop_requested
-        print("\nShutting down watch...")
-        stop_requested = True
-
-    signal.signal(signal.SIGINT, handle_signal)
-    signal.signal(signal.SIGTERM, handle_signal)
-
-    print("Starting polecat watch...")
-    print(f"  Polling interval: {interval}s")
-    print(f"  Stall threshold: {stall_threshold}min")
-    print(f"  Project filter: {project or 'all'}")
-    print("  Press Ctrl+C to stop.\n")
-
-    def _list(status, proj=project):
-        if manager.storage is not None and _use_legacy:
-            return manager.storage.list_tasks(
-                status=getattr(TaskStatus, status.upper()), project=proj
-            )
-        from polecat.pkb_bridge import list_tasks as pkb_list_tasks
-
-        return pkb_list_tasks(status=status, project=proj)
-
-    def _ready(proj=project):
-        if manager.storage is not None and _use_legacy:
-            return manager.storage.get_ready_tasks(project=proj)
-        from polecat.pkb_bridge import get_ready_tasks as pkb_ready
-
-        return pkb_ready(project=proj)
-
-    # Initial scan to populate seen sets (don't alert on startup)
-    try:
-        merge_ready_tasks = _list("merge_ready")
-        for task in merge_ready_tasks:
-            seen_merge_ready.add(task.id)
-
-        review_tasks = _list("review")
-        for task in review_tasks:
-            seen_review.add(task.id)
-
-        print(f"Initial state: {len(seen_merge_ready)} merge_ready, {len(seen_review)} review")
-    except Exception as e:
-        print(f"Warning: Initial scan failed: {e}")
-
-    while not stop_requested:
-        try:
-            now = datetime.now().astimezone()
-
-            # Check for new merge_ready tasks (new PRs filed)
-            merge_ready_tasks = _list("merge_ready")
-            for task in merge_ready_tasks:
-                if task.id not in seen_merge_ready:
-                    seen_merge_ready.add(task.id)
-                    last_activity = now
-                    _send_notification(
-                        "PR Filed",
-                        f"{task.id}: {task.title}",
-                        urgency="normal",
-                    )
-
-            # Check for new review tasks (merge failures)
-            review_tasks = _list("review")
-            for task in review_tasks:
-                if task.id not in seen_review:
-                    seen_review.add(task.id)
-                    last_activity = now
-                    _send_notification(
-                        "Review Needed",
-                        f"{task.id}: {task.title}",
-                        urgency="critical",
-                    )
-
-            # Check for completed tasks (mark as activity)
-            _list("done")
-            # We don't track done tasks, but finding new ones means progress
-            # This is a simplification - in production you'd track these too
-
-            # Check for in_progress tasks (active work)
-            in_progress = _list("in_progress")
-            if in_progress:
-                # Check if any were modified recently
-                for task in in_progress:
-                    task_mod = task.modified
-                    if task_mod and task_mod.tzinfo is None:
-                        task_mod = task_mod.replace(tzinfo=UTC)
-                    if task_mod and task_mod > last_activity:
-                        last_activity = task_mod
-
-            # Get leaf-ready tasks (actually pullable work)
-            leaf_ready = _ready()
-
-            # Check for stall
-            stall_cutoff = now - timedelta(minutes=stall_threshold)
-            if last_activity < stall_cutoff:
-                minutes_stalled = int((now - last_activity).total_seconds() / 60)
-                _send_notification(
-                    "Swarm Stalled",
-                    f"No progress in {minutes_stalled} minutes",
-                    urgency="critical",
-                )
-                # Reset to avoid spamming alerts
-                last_activity = now
-
-            # Status line (leaf-ready is the primary queue metric).
-            # Metric label uses the canonical PKB status `in_progress` (per
-            # aops-core/skills/remember/references/TAXONOMY.md) rather than
-            # the legacy "active" label.
-            ready_count = len(leaf_ready)
-            in_progress_count = len(in_progress)
-            merge_ready_count = len(merge_ready_tasks)
-            review_count = len(review_tasks)
-            timestamp = now.strftime("%H:%M:%S")
-            print(
-                f"[{timestamp}] ready={ready_count} in_progress={in_progress_count} merge_ready={merge_ready_count} review={review_count}"
-            )
-
-            # Record periodic metrics for dashboard
-            metrics.record_queue_depth("ready", count=ready_count, project=project)
-            metrics.record_queue_depth("in_progress", count=in_progress_count, project=project)
-            metrics.record_queue_depth("merge_ready", count=merge_ready_count, project=project)
-            metrics.record_queue_depth("review", count=review_count, project=project)
-
-        except Exception as e:
-            print(f"Error during poll: {e}")
-
-        # Sleep in small chunks to allow interrupt
-        for _ in range(interval):
-            if stop_requested:
-                break
-            time.sleep(1)
-
-    print("Watch stopped.")
+try:
+    from polecat.diagnostics import analyze as _analyze_cmd
+    from polecat.diagnostics import reset_stalled as _reset_stalled_cmd
+except ImportError:
+    from diagnostics import analyze as _analyze_cmd  # type: ignore[no-redef]
+    from diagnostics import reset_stalled as _reset_stalled_cmd  # type: ignore[no-redef]
+
+main.add_command(_analyze_cmd)
+main.add_command(_reset_stalled_cmd)
+
+
+try:
+    from polecat.watch import watch as _watch_cmd
+except ImportError:
+    from watch import watch as _watch_cmd  # type: ignore[no-redef]
+
+main.add_command(_watch_cmd)
 
 
 @main.command()
@@ -5295,280 +4772,12 @@ def swarm(ctx, claude, gemini, project, caller, dry_run):
     run_swarm(claude, gemini, project, caller, dry_run, str(home) if home else None)
 
 
-def parse_duration(duration_str: str) -> int:
-    """Parse a duration string like '8h', '1d', '30m' into seconds.
+try:
+    from polecat.summary import summary as _summary_cmd
+except ImportError:
+    from summary import summary as _summary_cmd  # type: ignore[no-redef]
 
-    Args:
-        duration_str: Duration string with suffix h (hours), d (days), or m (minutes)
-
-    Returns:
-        Duration in seconds
-
-    Raises:
-        ValueError: If format is invalid
-    """
-    if not duration_str:
-        raise ValueError("Duration string cannot be empty")
-
-    duration_str = duration_str.strip().lower()
-
-    # Handle numeric-only input (default to hours)
-    if duration_str.isdigit():
-        return int(duration_str) * 3600
-
-    if len(duration_str) < 2:
-        raise ValueError(f"Invalid duration format: {duration_str}")
-
-    value_str = duration_str[:-1]
-    unit = duration_str[-1]
-
-    try:
-        value = float(value_str)
-    except ValueError as e:
-        raise ValueError(f"Invalid duration value: {value_str}") from e
-
-    multipliers = {
-        "m": 60,  # minutes
-        "h": 3600,  # hours
-        "d": 86400,  # days
-    }
-
-    if unit not in multipliers:
-        raise ValueError(f"Unknown duration unit: {unit}. Use m, h, or d")
-
-    return int(value * multipliers[unit])
-
-
-@main.command()
-@click.option(
-    "--since",
-    "-s",
-    default="8h",
-    help="Time period to summarize (e.g., 8h, 1d, 30m). Default: 8h",
-)
-@click.option("--project", "-p", help="Filter by project (default: all)")
-@click.pass_context
-def summary(ctx, since, project):
-    """Generate a summary of polecat swarm work.
-
-    Shows merged PRs, completed tasks, and queue changes for the specified
-    time period. Output is markdown suitable for daily notes.
-
-    Examples:
-        polecat summary                # Last 8 hours
-        polecat summary --since 1d     # Last day
-        polecat summary -s 4h -p aops  # Last 4 hours, aops only
-    """
-    import json
-    import subprocess
-    from datetime import datetime, timedelta
-
-    manager = PolecatManager(home_dir=ctx.obj.get("home"))
-
-    def _list(status, proj=project):
-        if manager.storage is not None:
-            try:
-                from lib.task_model import TaskStatus
-
-                return manager.storage.list_tasks(
-                    status=getattr(TaskStatus, status.upper()), project=proj
-                )
-            except ImportError:
-                pass
-        from polecat.pkb_bridge import list_tasks as pkb_list_tasks
-
-        return pkb_list_tasks(status=status, project=proj)
-
-    def _ready(proj=project):
-        if manager.storage is not None:
-            try:
-                from lib.task_model import TaskStatus  # noqa: F401, F811
-
-                return manager.storage.get_ready_tasks(project=proj)
-            except ImportError:
-                pass
-        from polecat.pkb_bridge import get_ready_tasks as pkb_ready
-
-        return pkb_ready(project=proj)
-
-    # Parse the duration
-    try:
-        seconds = parse_duration(since)
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    cutoff = datetime.now().astimezone() - timedelta(seconds=seconds)
-    cutoff_iso = cutoff.strftime("%Y-%m-%dT%H:%M:%S")
-
-    # Format duration for display
-    if seconds >= 86400:
-        duration_display = f"{seconds // 86400} day(s)"
-    elif seconds >= 3600:
-        duration_display = f"{seconds // 3600} hour(s)"
-    else:
-        duration_display = f"{seconds // 60} minute(s)"
-
-    print(f"## Polecat Swarm Summary (last {duration_display})")
-    print()
-
-    # --- Merged PRs ---
-    print("### PRs Merged")
-    print()
-
-    merged_prs = []
-    try:
-        # Query GitHub for merged PRs
-        # gh pr list --state merged returns PRs merged, filtered by date
-        gh_cmd = [
-            "gh",
-            "pr",
-            "list",
-            "--state",
-            "merged",
-            "--search",
-            f"merged:>{cutoff_iso[:10]}",  # Date only for search
-            "--json",
-            "number,title,mergedAt,headRefName",
-            "--limit",
-            "100",
-        ]
-
-        result = subprocess.run(gh_cmd, capture_output=True, text=True, check=False)
-
-        if result.returncode == 0 and result.stdout.strip():
-            all_prs = json.loads(result.stdout)
-
-            # Filter by actual merged time (gh search is date-only)
-            for pr in all_prs:
-                merged_at = pr.get("mergedAt", "")
-                if merged_at:
-                    # Parse ISO timestamp
-                    try:
-                        merged_dt = datetime.fromisoformat(merged_at.replace("Z", "+00:00"))
-                        if merged_dt >= cutoff:
-                            # Filter by project if specified
-                            if project:
-                                # Check if branch matches project pattern
-                                branch = pr.get("headRefName", "")
-                                if not branch.startswith("polecat/"):
-                                    continue
-                            merged_prs.append(pr)
-                    except (ValueError, TypeError):
-                        pass
-
-            if merged_prs:
-                print(f"**{len(merged_prs)} PRs merged**")
-                print()
-                for pr in merged_prs[:20]:  # Limit display
-                    print(f"- #{pr['number']}: {pr['title']}")
-                if len(merged_prs) > 20:
-                    print(f"- ... and {len(merged_prs) - 20} more")
-            else:
-                print("No PRs merged in this period.")
-        else:
-            print("(Could not query GitHub - gh CLI not available or not authenticated)")
-
-    except FileNotFoundError:
-        print("(GitHub CLI not installed)")
-    except Exception as e:
-        print(f"(GitHub query failed: {e})")
-
-    print()
-
-    # --- Completed Tasks ---
-    print("### Tasks Completed")
-    print()
-
-    completed_tasks = []
-    try:
-        # Get all done tasks and filter by modified time
-        all_done = _list("done")
-
-        for task in all_done:
-            task_mod = task.modified
-            if task_mod is None:
-                continue
-            # Handle both date and datetime objects
-            if hasattr(task_mod, "tzinfo"):
-                # It's a datetime
-                if task_mod.tzinfo is None:
-                    task_mod = task_mod.replace(tzinfo=UTC)
-            else:
-                # It's a date - convert to datetime at midnight UTC
-                task_mod = datetime.combine(task_mod, datetime.min.time(), tzinfo=UTC)
-
-            if task_mod >= cutoff:
-                completed_tasks.append(task)
-
-        if completed_tasks:
-            print(f"**{len(completed_tasks)} tasks completed**")
-            print()
-            for task in completed_tasks[:20]:
-                print(f"- [{task.id}] {task.title}")
-            if len(completed_tasks) > 20:
-                print(f"- ... and {len(completed_tasks) - 20} more")
-        else:
-            print("No tasks completed in this period.")
-
-    except Exception as e:
-        print(f"(Task query failed: {e})")
-
-    print()
-
-    # --- Queue Status ---
-    print("### Queue Status")
-    print()
-
-    try:
-        # Count tasks by status
-        ready_tasks = _ready()
-        in_progress = _list("in_progress")
-        blocked = _list("blocked")
-        review = _list("review")
-        merge_ready = _list("merge_ready")
-
-        print(f"- **Ready**: {len(ready_tasks)} tasks")
-        print(f"- **In Progress**: {len(in_progress)} tasks")
-        print(f"- **Blocked**: {len(blocked)} tasks")
-        print(f"- **Review**: {len(review)} tasks")
-        print(f"- **Merge Ready**: {len(merge_ready)} tasks")
-        print(f"- **Completed** (this period): {len(completed_tasks)} tasks")
-
-    except Exception as e:
-        print(f"(Queue status query failed: {e})")
-
-    print()
-
-    # --- Active Workers ---
-    print("### Active Workers")
-    print()
-
-    try:
-        # Count active polecats (worktrees)
-        active_polecats = [
-            d.name
-            for d in manager.polecats_dir.iterdir()
-            if d.is_dir() and not d.name.startswith(".")
-        ]
-
-        # Count crew workers
-        crew_workers = manager.list_crew()
-
-        if active_polecats:
-            print(f"- **Polecats**: {len(active_polecats)} active worktrees")
-        else:
-            print("- **Polecats**: None active")
-
-        if crew_workers:
-            print(f"- **Crew**: {len(crew_workers)} workers ({', '.join(crew_workers)})")
-        else:
-            print("- **Crew**: None active")
-
-    except Exception as e:
-        print(f"(Worker status query failed: {e})")
-
-    print()
+main.add_command(_summary_cmd)
 
 
 if __name__ == "__main__":
